requirements.txt
fastapi==0.78.0
python-dotenv==0.20.0
requests==2.28.0
uvicorn==0.17.6
gql[all]